chore: Update 'bookmarks' app to have stricter public API
